스테이지 2개 추가

+++ David Engine Project/Engine Source/daMath.cpp
@@ -255,8 +255,8 @@ namespace da::math
 	Vector2 daRotateVector2(const Vector2& vector, float radian) noexcept
 	{
 		Vector2 retVector;
-		double cosA = std::cos(radian);
-		double sinA = std::sin(radian);
+		float cosA = std::cos(radian);
+		float sinA = std::sin(radian);
 
 		retVector.x = vector.x * cosA - vector.y * sinA;
 		retVector.y = vector.x * sinA + vector.y * cosA;
